# 8306308: (ch) Writer created by Channels::newWriter may lose data