--- a/src/java.base/share/classes/java/nio/channels/Channels.java
+++ b/src/java.base/share/classes/java/nio/channels/Channels.java
@@ -29,7 +29,6 @@ import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.io.OutputStreamWriter;
 import java.io.Reader;
 import java.io.Writer;
 import java.io.IOException;
@@ -542,7 +541,9 @@ public final class Channels {
                                    int minBufferCap)
     {
         Objects.requireNonNull(ch, "ch");
-        return new OutputStreamWriter(newOutputStream(ch), enc.reset());
+        Objects.requireNonNull(enc, "enc");
+        return StreamEncoder.forOutputStreamWriter(newOutputStream(ch),
+            new Object(), enc.reset());
     }
 
     /**