--- a/src/java.base/share/classes/sun/nio/cs/StreamEncoder.java
+++ b/src/java.base/share/classes/sun/nio/cs/StreamEncoder.java
@@ -31,7 +31,6 @@ import java.io.UnsupportedEncodingException;
 import java.io.Writer;
 import java.nio.ByteBuffer;
 import java.nio.CharBuffer;
-import java.nio.channels.WritableByteChannel;
 import java.nio.charset.Charset;
 import java.nio.charset.CharsetEncoder;
 import java.nio.charset.CoderResult;
@@ -78,17 +77,6 @@ public final class StreamEncoder extends Writer {
     {
         return new StreamEncoder(out, lock, enc);
     }
-
-
-    // Factory for java.nio.channels.Channels.newWriter
-
-    public static StreamEncoder forEncoder(WritableByteChannel ch,
-                                           CharsetEncoder enc,
-                                           int minBufferCap)
-    {
-        return new StreamEncoder(ch, enc, minBufferCap);
-    }
-
 
     // -- Public methods corresponding to those in OutputStreamWriter --
 
@@ -252,9 +240,8 @@ public final class StreamEncoder extends Writer {
     private ByteBuffer bb;
     private final int maxBufferCapacity;
 
-    // Exactly one of these is non-null
+    // This must be non-null
     private final OutputStream out;
-    private final WritableByteChannel ch;
 
     // Leftover first char in a surrogate pair
     private boolean haveLeftoverChar = false;
@@ -271,27 +258,11 @@ public final class StreamEncoder extends Writer {
     private StreamEncoder(OutputStream out, Object lock, CharsetEncoder enc) {
         super(lock);
         this.out = out;
-        this.ch = null;
         this.cs = enc.charset();
         this.encoder = enc;
 
         this.bb = ByteBuffer.allocate(INITIAL_BYTE_BUFFER_CAPACITY);
         this.maxBufferCapacity = MAX_BYTE_BUFFER_CAPACITY;
-    }
-
-    private StreamEncoder(WritableByteChannel ch, CharsetEncoder enc, int mbc) {
-        this.out = null;
-        this.ch = ch;
-        this.cs = enc.charset();
-        this.encoder = enc;
-
-        if (mbc > 0) {
-            this.bb = ByteBuffer.allocate(mbc);
-            this.maxBufferCapacity = mbc;
-        } else {
-            this.bb = ByteBuffer.allocate(INITIAL_BYTE_BUFFER_CAPACITY);
-            this.maxBufferCapacity = MAX_BYTE_BUFFER_CAPACITY;
-        }
     }
 
     private void writeBytes() throws IOException {
@@ -302,12 +273,7 @@ public final class StreamEncoder extends Writer {
         int rem = (pos <= lim ? lim - pos : 0);
 
         if (rem > 0) {
-            if (ch != null) {
-                int wc = ch.write(bb);
-                assert wc == rem : rem;
-            } else {
-                out.write(bb.array(), bb.arrayOffset() + pos, rem);
-            }
+            out.write(bb.array(), bb.arrayOffset() + pos, rem);
         }
         bb.clear();
     }
@@ -414,7 +380,7 @@ public final class StreamEncoder extends Writer {
     }
 
     void implClose() throws IOException {
-        try (ch; out) {
+        try (out) {
             flushLeftoverChar(null, true);
             for (;;) {
                 CoderResult cr = encoder.flush(bb);