@bplb 2023-10-16
I am not entirely certain that this change is exactly what @djelinski intended in his [comment](https://bugs.openjdk.org/browse/JDK-8306308?focusedCommentId=14575246&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-14575246) on the issue.

No tests are included pending comments on whether this is the right approach.