@AlanBateman 2023-10-17
It's a simple approach but I think you'll create to some some benchmarks to demonstrate that the channel handling in StreamEncoder can be removed.