@bplb 2023-10-17
This was just a first stab to elicit comments and verify the approach. Note that this would make `minBufferCap` irrelevant. I wonder whether some sort of `@xyzNote` would be appropriate to make that obvious.