@bplb 2023-10-18
> [...] so this PR has potential to make the code much simpler.

So simplified in 1d8a667c0160fe9d698773e6bbf8abc2b8ee2596.