@bplb 2023-10-18
> [...] you'll need to create to some benchmarks [...]

Problems running benchmarks today, likely due to version considerations. Will continue that and if it pans out add tests.