@AlanBateman 2023-10-19
BTW: Do we have tests with SocketChannels configured non-blocking that will check that IllegalBlockingMode is thrown?