@djelinski 2023-10-19
This is benchmarking disk access. Use a blackhole channel instead:
```
    public void create(Blackhole bh) throws IOException {
        cbuf = new char[capacity];
        Arrays.fill(cbuf, (char)42);
        WritableByteChannel fc = new WritableByteChannel() {
            @Override
            public int write(ByteBuffer src) throws IOException {
                bh.consume(src);
                int remaining = src.remaining();
                src.position(src.limit());
                return remaining;
            }
            @Override
            public boolean isOpen() {
                return true;
            }
            @Override
            public void close() throws IOException {
            }
        };
        writer = Channels.newWriter(fc, "UTF-8");
    }
```
Before:
```
Benchmark               (capacity)   Mode  Cnt     Score     Error   Units
WriterBench.testMethod        4096  thrpt   30  7679,273 � 111,851  ops/ms
WriterBench.testMethod        8192  thrpt   30  4194,707 �  47,489  ops/ms
WriterBench.testMethod       16384  thrpt   30  2191,705 �  46,152  ops/ms
WriterBench.testMethod       32768  thrpt   30  1040,875 �  54,422  ops/ms
```
After:
```
Benchmark               (capacity)   Mode  Cnt     Score     Error   Units
WriterBench.testMethod        4096  thrpt   30  7414,422 � 156,937  ops/ms
WriterBench.testMethod        8192  thrpt   30  4161,465 � 116,885  ops/ms
WriterBench.testMethod       16384  thrpt   30  2387,347 �  37,324  ops/ms
WriterBench.testMethod       32768  thrpt   30  1168,550 �  35,374  ops/ms
```
(well, still no difference)