@djelinski 2023-10-19
> Do we have tests with SocketChannels configured non-blocking that will check that IllegalBlockingMode is thrown?

I don't believe we do; the code attached to the JBS ticket suggests that we did not throw, so if we had such a test, it would be failing.