@bplb 2023-10-19
@AlanBateman and @djelinski thanks for all the comments. I will address them and add test coverage, which I inteentionally deferred.