@bplb 2023-10-19
> (well, still no difference)

Likewise.

Before:
```
Benchmark               (capacity)   Mode  Cnt        Score        Error  Units
WriterBench.testMethod        4096  thrpt    5  7896495.475 ± 109101.117  ops/s
WriterBench.testMethod        8192  thrpt    5  3507455.124 ± 125380.018  ops/s
WriterBench.testMethod       16384  thrpt    5  1761486.228 ±  57484.438  ops/s
WriterBench.testMethod       32768  thrpt    5   897733.089 ±  88233.815  ops/s
```
After:
```
Benchmark               (capacity)   Mode  Cnt        Score        Error  Units
WriterBench.testMethod        4096  thrpt    5  7383045.614 ±  72860.403  ops/s
WriterBench.testMethod        8192  thrpt    5  4073733.378 ±  30800.651  ops/s
WriterBench.testMethod       16384  thrpt    5  1789169.956 ± 183912.728  ops/s
WriterBench.testMethod       32768  thrpt    5   848052.004 ± 113203.131  ops/s
```
I don't know why the measurements have such a large numerical difference however.