@bplb 2023-10-19
> I don't know why the measurements have such a large numerical difference however.

D'oh: `ops/s` instead of `ops/ms`.