@bplb 2023-10-19
> BTW: Do we have tests with SocketChannels configured non-blocking that will check that IllegalBlockingMode is thrown?

The test I just added should cover this.