@bplb 2023-10-19
The test I added is just the reproducers attached by @djelinski (thanks!) to the JDK-8306308 issue aggregated into a single JUnit 5 test. Both sub-tests fail when run against the master branch, but pass with the proposed change applied.